update: refactoring code
