+++ README.md
@@ -27,12 +27,11 @@
 
 ## Version Release
 
-    $version_release = 5.0.2
+    $version_release = 5.1.0
 
     // Suport Library
     $admob_version = 21.1.0
     $unity_ad_version = 4.2.1
-    $start_io_version = 4.10.0
 
 What's New??
 
@@ -89,23 +88,17 @@ allprojects {
         // library unity ads (Required)
         implementation 'com.unity3d.ads:unity-ads:${unity_ad_version}'
 
-        // library start.io (Required)
-        implementation 'com.startapp:inapp-sdk:${$start_io_version}'
-
         // library frogo-admob (Required - Recomended)
-        implementation 'com.github.amirisback:frogo-admob:5.0.2'
+        implementation 'com.github.amirisback:frogo-admob:5.1.0'
 
         // -----------------------------------------------------------------------------------------
         // For Single Library Patch 
 
         // library frogo-admob (Admob Only)
-        implementation 'com.github.amirisback.frogo-admob:frogoadmob:5.0.2'
+        implementation 'com.github.amirisback.frogo-admob:frogoadmob:5.1.0'
 
         // library frogo-admob (Unity Ads Only)
-        implementation 'com.github.amirisback.frogo-admob:frogounityad:5.0.2'
-
-        // library frogo-admob (Start Io Only)
-        implementation 'com.github.amirisback.frogo-admob:frogostartioad:5.0.2'
+        implementation 'com.github.amirisback.frogo-admob:frogounityad:5.1.0'
 	}
 
 #### <Option 2> Kotlin DSL
@@ -117,23 +110,17 @@ allprojects {
         // library unity ads (Required)
         implementation("com.unity3d.ads:unity-ads:${unity_ad_version}")
 
-        // library start.io (Required)
-        implementation("com.startapp:inapp-sdk:${$start_io_version")
-
         // library frogo-admob (Required - Recomended)
-        implementation("com.github.amirisback:frogo-admob:5.0.2")
+        implementation("com.github.amirisback:frogo-admob:5.1.0")
 
         // -----------------------------------------------------------------------------------------
         // For Single Library Patch
 
         // library frogo-admob (Admob Only)
-        implementation("com.github.amirisback.frogo-admob:frogoadmob:5.0.2")
+        implementation("com.github.amirisback.frogo-admob:frogoadmob:5.1.0")
 
         // library frogo-admob (Unity Ads Only)
-        implementation("com.github.amirisback.frogo-admob:frogounityad:5.0.2")
-
-        // library frogo-admob (Start Io Only)
-        implementation("com.github.amirisback.frogo-admob:frogostartioad:5.0.2")
+        implementation("com.github.amirisback.frogo-admob:frogounityad:5.1.0")
 
 	}
