+++ buildSrc/src/main/kotlin/DependencyGradle.kt
@@ -18,9 +18,9 @@ object DependencyGradle {
     const val COMPOSE_VERSION = Version.Androidx.compose
 
     const val FROGO_UI_VERSION = "1.0.8"
-    const val FROGO_SDK_VERSION = "2.0.5"
-    const val FROGO_CONSUME_API_VERSION = "2.3.7"
-    const val FROGO_RECYCLER_VIEW_VERSION = "4.2.3"
+    const val FROGO_SDK_VERSION = "2.0.6"
+    const val FROGO_CONSUME_API_VERSION = "2.3.8"
+    const val FROGO_RECYCLER_VIEW_VERSION = "4.3.0"
 
     const val FROGO_UI = "com.github.frogobox:frogo-ui:$FROGO_UI_VERSION"
     const val FROGO_SDK = "com.github.frogobox:frogo-sdk:$FROGO_SDK_VERSION"
