+++ buildSrc/src/main/kotlin/ProjectSetting.kt
@@ -23,8 +23,8 @@ object ProjectSetting {
     // ---------------------------------------------------------------------------------------------
 
     const val VERSION_MAJOR = 5
-    const val VERSION_MINOR = 0
-    const val VERSION_PATCH = 2
+    const val VERSION_MINOR = 1
+    const val VERSION_PATCH = 0
 
     // ---------------------------------------------------------------------------------------------
