+++ frogoadmob/src/main/java/com/frogobox/admob/core/IFrogoAdmobInterstitial.kt
@@ -1,0 +1,79 @@
+package com.frogobox.admob.core
+
+import androidx.appcompat.app.AppCompatActivity
+import com.frogobox.admob.callback.FrogoAdmobInterstitialCallback
+
+/**
+ * Created by Faisal Amir
+ * FrogoBox Inc License
+ * =========================================
+ * ImplementationAdmob
+ * Copyright (C) 10/02/2020.
+ * All rights reserved
+ * -----------------------------------------
+ * Name     : Muhammad Faisal Amir
+ * E-mail   : faisalamircs@gmail.com
+ * Github   : github.com/amirisback
+ * LinkedIn : linkedin.com/in/faisalamircs
+ * -----------------------------------------
+ * FrogoBox Software Industries
+ * com.frogobox.admob
+ *
+ */
+
+
+interface IFrogoAdmobInterstitial {
+
+    fun showAdInterstitial(
+        activity: AppCompatActivity,
+        interstitialAdUnitId: String,
+        timeoutMilliSecond: Int?,
+        keyword: List<String>?,
+        callback: FrogoAdmobInterstitialCallback?
+    )
+
+    fun showAdInterstitial(
+        activity: AppCompatActivity,
+        interstitialAdUnitId: String,
+        timeoutMilliSecond: Int,
+        keyword: List<String>
+    )
+
+    fun showAdInterstitial(
+        activity: AppCompatActivity,
+        interstitialAdUnitId: String,
+        timeoutMilliSecond: Int
+    )
+
+    fun showAdInterstitial(
+        activity: AppCompatActivity,
+        interstitialAdUnitId: String,
+        keyword: List<String>
+    )
+
+    fun showAdInterstitial(
+        activity: AppCompatActivity,
+        interstitialAdUnitId: String,
+    )
+
+    fun showAdInterstitial(
+        activity: AppCompatActivity,
+        interstitialAdUnitId: String,
+        timeoutMilliSecond: Int,
+        callback: FrogoAdmobInterstitialCallback
+    )
+
+    fun showAdInterstitial(
+        activity: AppCompatActivity,
+        interstitialAdUnitId: String,
+        keyword: List<String>,
+        callback: FrogoAdmobInterstitialCallback
+    )
+
+    fun showAdInterstitial(
+        activity: AppCompatActivity,
+        interstitialAdUnitId: String,
+        callback: FrogoAdmobInterstitialCallback
+    )
+
+}
